+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.5.0] - 13.11.2023
+
 ### Added
 
 - Added Vietnamese
@@ -202,7 +204,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 <br>
 
-[unreleased]: https://github.com/an-anime-team/the-honkers-railway-launcher/compare/1.4.0...next
+[unreleased]: https://github.com/an-anime-team/the-honkers-railway-launcher/compare/1.5.0...next
+[1.5.0]: https://github.com/an-anime-team/the-honkers-railway-launcher/compare/1.4.0...1.5.0
 [1.4.0]: https://github.com/an-anime-team/the-honkers-railway-launcher/compare/1.3.0...1.4.0
 [1.3.0]: https://github.com/an-anime-team/the-honkers-railway-launcher/compare/1.2.5...1.3.0
 [1.2.5]: https://github.com/an-anime-team/the-honkers-railway-launcher/compare/1.2.4...1.2.5